[Feat] #11 HealthUtility.swift > HealthStore.swift 파일명 변경, HealthStore 싱글톤 패턴으로 리팩토링
